+++ fft.h
@@ -1,5 +1,6 @@
 #ifndef ARDUINO_FFT_INCLUDED
 #define ARDUINO_FFT_INCLUDED
+
 // An FFT for the Arduino.
 // This was built on UNIX the C declarations
 // and C code in fft.c can be placed into
@@ -8,9 +9,9 @@
 // The arrays for the fft will be computed in place
 // and thus your array will have the fft result
 // written over your original data
-void fft(double data_re[], double data_im[],const int N);
+void fft(float data_re[], float data_im[],const int N);
 // helper functions called by the fft
 // data will first be rearranged then computed
-void rearrange(double data_re[],double data_im[],const unsigned int N);
-void compute(double data_re[],double data_im[],const unsigned int N);
+void rearrange(float data_re[],float data_im[],const unsigned int N);
+void compute(float data_re[],float data_im[],const unsigned int N);
 #endif
